Fix coldfront processor
During this month’s invoicing, several bugs were found with the Coldfront processor:

- The test case function `_get_mock_allocation_data` returned
an incorrect representation of the Coldfront API data. This meant
the test cases passed, but the code would crash when calling the real API.
- The “giant query URL” method led an error returned from
the actual API server. For now, we’ll use `?all=true` to return
Coldfront allocations instead. This should be replaced with
a filter in the future.
- The processor incorrectly assumed that Project IDs were
the same as Project Names in all invoices. While this *is* true
for Openshift, it is not for other services.

In addition to fixing these bugs, the following changes have
been added to increase confidence in this processor:
- If a project ID is not found in Coldfront, and it is also
billable, an error will be raised instead of just a log message.
- When extracting values from Coldfront API, ignore allocations
if any field except institute code is not set.