Added nonbillable cluster list to billable processor
As we plan to process projects from the non-billable `ocp-test` cluster [1],
a list is added containing just `ocp-test`. This list is also used by the
Coldfront processor when validating the Coldfront API data.
As processing now has cluster-specific processing, a new
column `Cluster Name` is added.
Test cases have been updated accordingly.

While revising test cases, two bugs were found:
- In `TestColdfrontFetchProcessor._get_test_invoice`, the
default value for project names was set to be a copy of project ids.
This didn’t invalidate pre-existing test cases, but would have
invalidated the test that checked for nonbillable clusters.
- in `TestValidateBillablePIProcessor.test_remove_nonbillables`,
in the last assertion which checked if the output invoice only
contained billable PIs, the assertion only checked if the invoice
contained PIs from the list, not if the invoice ALL PIs from it, and
with an exact quantity. This bug does invalidate the test.
Fortunately, test cases still passed after the bug is fixed.

[1] https://github.com/CCI-MOC/invoicing/issues/167