Added CLI option to pass in Coldfront API data
The filepath to a JSON file can be now passed through
the CLI. This will allow injection of custom Coldfront
data for ease of testing purposes.

`validate_required_env_vars` has been changed. It now
takes in a list of env var names to check. Keycloak credentials
will only be checked if a Coldfront API data file is not provided.