Make chromium test expect configured chrome binary path
The chrome binary used for pdf generation is controlled by the
CHROME_BIN_PATH environment variable, but the test in in
process_report/tests/integration/test_chromium.py expected the literal path
`/usr/bin/chromium`. This would cause the tests to fail if run in an
environment in which chromium was not installed, even if the code
successfully used an alternate path.